# frontend/packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateCustomFoodRequest": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "dietary_class": {
            "$ref": "#/components/schemas/DietaryClass"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "category_id",
          "dietary_class"
        ],
        "title": "CreateCustomFoodRequest",
        "type": "object"
      },
...
      "DietaryClass": {
        "description": "What the person eats — DB §8.5, FR-M4-035.\n\n🔒 Lives on the *client*, not the plan (DB §5.1): it is a property of the\nperson, and it filters food search at authoring time, before a plan exists.\n\n``JAIN`` is not a stricter vegetarian in a way a boolean could express — it\nexcludes root vegetables, which is a different axis from animal products\nentirely, and collapsing it would put onion in a Jain client's plan.",
        "enum": [
          "vegetarian",
          "eggetarian",
          "non_vegetarian",
          "vegan",
          "jain"
        ],
        "title": "DietaryClass",
        "type": "string"
      },
...
      "FoodItemResponse": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "dietary_class": {
            "$ref": "#/components/schemas/DietaryClass"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "required": [
          "id",
          "name",
          "category_id",
          "dietary_class",
          "tenant_id"
        ],
        "title": "FoodItemResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Practice management for Indian nutrition practitioners.\n\n🔒 Realm-segmented paths (ADR-A01): `/app` practitioner, `/portal` client, `/admin` operator, `/public` unauthenticated.",
    "title": "WellnessCRM V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app/nutrition/foods": {
      "get": {
        "description": "Search the food catalogue (FR-M4-008, AC-M4-003, AC-M4-009).",
        "operationId": "searchFoods",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dietary_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DietaryClass"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dietary Class"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FoodItemResponse"
                  },
                  "title": "Response Searchfoods",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Foods",
        "tags": [
          "nutrition-catalogue"
        ]
      },
      "post": {
        "description": "Create a new custom food for this tenant (FR-M4-012, FR-M4-013).",
        "operationId": "createCustomFood",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomFoodRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Custom Food",
        "tags": [
          "nutrition-catalogue"
        ]
      }
    },
...
  }
}